+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Implemented `LossPlotter._print_error_stats` to report loss breakdown of error terms.
 - Added `FDNetwork.is_edge_fully_supported`. 
 
 ### Changed